move Instance to bokeh.core.property.instance
